pyproject.toml
[project]
name = "showcase"
version = "0.1.0"
description = "Resume processing pipeline with OCR, AI, and frontend generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "google-genai>=0.2.0",
    "aiofiles>=23.2.0",
    "sqlmodel>=0.0.31",
    "httpx>=0.28.1",
    "email-validator>=2.3.0",
    "tenacity>=9.1.2",
    "firebase-admin>=7.1.0",
    "jose>=1.0.0",
    "python-jose[cryptography]>=3.5.0",
    "agno>=2.3.24",
    "openai>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
    "httpx>=0.28.1",  # For TestClient
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "agents", "setup_scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise from in except handler
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in init files
"tests/*" = ["B011"]      # assert false allowed in tests

[tool.black]
line-length = 100
target-version = ["py311"]

